Fix threshold config variable

+++ image_object_detection/detection/config.py
@@ -2,4 +2,4 @@ import os
 
 
 DETECTION_CLASSES = os.environ.get('DETECTION_CLASSES', 'person').split(',')
-DETECTION_THRESHOLD = int(os.environ.get('DETECTION_THRESHOLD') or 0.5)
+DETECTION_THRESHOLD = float(os.environ.get('DETECTION_THRESHOLD') or 0.5)
